+++ HISTORY.rst
@@ -1,5 +1,17 @@
 Bug tracker at https://github.com/giampaolo/pyftpdlib/issues
 
+Version: 1.5.5 - 2019-04-04
+===========================
+
+**Enhancements**
+
+- #495: colored test output.
+
+**Bug fixes**
+
+- #492: CRLF line endings are replaced with CRCRLF in ASCII mode downloads.
+- #496: import error due to multiprocessing.Lock() bug.
+
 Version: 1.5.4 - 2018-05-04
 ===========================